style: graphic refont adapted to figma
